fix helm values using correct make file arguments
