+++ .nvim/nvim-dap.lua
@@ -1,15 +1,5 @@
 local dap = require('dap')
 
-dap.adapters.cilium_delve = {
-  type = 'server',
-  port = '${port}',
-  executable = {
-    command = 'dlv',
-    args = {'dap', '-l', '127.0.0.1:${port}'},
-    -- add this if on windows, otherwise server won't open successfully
-    -- detached = false
-  }
-}
 dap.adapters.cilium_kind_control_plane_1 = {
     type = "server",
     host = "127.0.0.1",
@@ -31,21 +21,6 @@ dap.adapters.cilium_operator_kind_worker_1 = {
     port = 23511
 }
 dap.configurations.go = {
-    {
-        name = "Debug unit tests in the current file",
-        type = "cilium_delve",
-        request = "launch",
-        mode = "test",
-        program = "./${relativeFileDirname}",
-    },
-    {
-        name = "Debug controlplane test (open test/controlplane/${test}/*.go first)",
-        type = "cilium_delve",
-        request = "launch",
-        mode = "test",
-        program = "./${relativeFileDirname}/../",
-        args = { "-test.v", "-test.run", "TestControlPlane/${fileDirname}"},
-    },
     {
         type = "cilium_kind_control_plane_1",
         request = "attach",
