+++ .vscode/launch.json
@@ -12,14 +12,6 @@
 			"program": "${relativeFileDirname}",
 		},
 		{
-			"name": "Launch controlplane test",
-			"type": "go",
-			"request": "launch",
-			"mode": "test",
-			"program": "${relativeFileDirname}/../",
-			"args": ["-test.v", "-test.run", "TestControlPlane/${fileDirname}"],
-		},
-		{
 			"name": "Attach to kind-control-plane-1",
 			"type": "go",
 			"request": "attach",
