+++ daemon/cmd/cells.go
@@ -34,10 +34,8 @@ import (
 	"github.com/cilium/cilium/pkg/endpointmanager"
 	"github.com/cilium/cilium/pkg/envoy"
 	"github.com/cilium/cilium/pkg/gops"
-	identity "github.com/cilium/cilium/pkg/identity/cache/cell"
 	"github.com/cilium/cilium/pkg/identity/identitymanager"
 	ipamcell "github.com/cilium/cilium/pkg/ipam/cell"
-	ipcache "github.com/cilium/cilium/pkg/ipcache/cell"
 	"github.com/cilium/cilium/pkg/k8s"
 	k8sClient "github.com/cilium/cilium/pkg/k8s/client"
 	k8sSynced "github.com/cilium/cilium/pkg/k8s/synced"
@@ -54,7 +52,6 @@ import (
 	nodeManager "github.com/cilium/cilium/pkg/node/manager"
 	"github.com/cilium/cilium/pkg/nodediscovery"
 	"github.com/cilium/cilium/pkg/option"
-	policy "github.com/cilium/cilium/pkg/policy/cell"
 	policyDirectory "github.com/cilium/cilium/pkg/policy/directory"
 	policyK8s "github.com/cilium/cilium/pkg/policy/k8s"
 	"github.com/cilium/cilium/pkg/pprof"
@@ -223,11 +220,17 @@ var (
 		// Auth is responsible for authenticating a request if required by a policy.
 		auth.Cell,
 
-		// Provides IdentityAllocators (Responsible for allocating security identities)
-		identity.Cell,
+		// Provides the IPCache
+		cell.Provide(newIPCache),
 
-		// IPCache cell provides IPCache (IP to identity mappings)
-		ipcache.Cell,
+		// Provides PolicyRepository
+		cell.Provide(newPolicyRepo),
+
+		// Provides PolicyUpdater
+		cell.Provide(newPolicyUpdater),
+
+		// Provides the different types of IdentityAllocators
+		cell.Provide(newIdentityAllocator),
 
 		// IPAM provides IP address management.
 		ipamcell.Cell,
@@ -237,9 +240,6 @@ var (
 
 		// ServiceCache holds the list of known services correlated with the matching endpoints.
 		k8s.ServiceCacheCell,
-
-		// Provides PolicyRepository (List of policy rules)
-		policy.Cell,
 
 		// K8s policy resource watcher cell. It depends on the half-initialized daemon which is
 		// resolved by newDaemonPromise()
