+++ install/kubernetes/cilium/values.yaml.tmpl
@@ -1811,6 +1811,7 @@ hubble:
   # -- Emit v1.Events related to pods on detection of packet drops.
   #    This feature is alpha, please provide feedback at https://github.com/cilium/cilium/issues/33975.
   dropEventEmitter:
+    # -- Enable dropEventEmitter
     enabled: false
     # --- Minimum time between emitting same events.
     interval: 2m
